requirements.txt
fastapi 
uvicorn 
pydantic 
python-dotenv 
httpx 
crewai 
crewai-tools